refactor: improves auth codebase
